Fix ambiguous call to Value constructor

+++ src/vm/systemCalls.maps.cpp
@@ -231,25 +231,25 @@ void systemCallMapLen(const SystemCallInput& input, SystemCallResult* result) {
     switch (mapObject.getObjectType()) {
         case ObjectType::kObjectToObjectMap: {
             const auto& map = castObjectToObjectMap(mapObject);
-            result->returnedValue = Value{ map.pairs.size() };
+            result->returnedValue = Value{ static_cast<uint64_t>(map.pairs.size()) };
             break;
         }
 
         case ObjectType::kObjectToValueMap: {
             const auto& map = castObjectToValueMap(mapObject);
-            result->returnedValue = Value{ map.pairs.size() };
+            result->returnedValue = Value{ static_cast<uint64_t>(map.pairs.size()) };
             break;
         }
 
         case ObjectType::kValueToObjectMap: {
             const auto& map = castValueToObjectMap(mapObject);
-            result->returnedValue = Value{ map.pairs.size() };
+            result->returnedValue = Value{ static_cast<uint64_t>(map.pairs.size()) };
             break;
         }
 
         case ObjectType::kValueToValueMap: {
             const auto& map = castValueToValueMap(mapObject);
-            result->returnedValue = Value{ map.pairs.size() };
+            result->returnedValue = Value{ static_cast<uint64_t>(map.pairs.size()) };
             break;
         }
